chore: move isCSSRequest to utils
