+++ packages/vite/src/node/optimizer/esbuildDepPlugin.ts
@@ -6,12 +6,13 @@ import {
   escapeRegex,
   flattenId,
   isBuiltin,
+  isCSSRequest,
   isExternalUrl,
   moduleListContains,
   normalizePath,
 } from '../utils'
 import { browserExternalId, optionalPeerDepId } from '../plugins/resolve'
-import { isCSSRequest, isModuleCSSRequest } from '../plugins/css'
+import { isModuleCSSRequest } from '../plugins/css'
 import type { Environment } from '../environment'
 import { createBackCompatIdResolver } from '../idResolver'
